requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
astunparse==1.6.3
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cozmo==1.4.10
cozmoclad==3.4.0
cycler==0.10.0
Cython==3.0a5
decorator==4.4.2
distlib==0.3.0
filelock==3.0.12
Flask==1.1.2
gast==0.2.2
google-auth==1.15.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.29.0
h5py==2.10.0
idna==2.9
imagecodecs==2020.2.18
imageio==2.8.0
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.1
networkx==2.4
numpy==1.16.6
oauthlib==3.1.0
object-detection==0.1
opencv-python==4.2.0.34
opt-einsum==3.2.1
pandas==1.0.3
pbr==5.4.5
Pillow==7.1.2
protobuf==3.12.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.0
pyparsing==2.4.7
PyQt5==5.14.2
PyQt5-sip==12.7.2
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.17.2
scikit-learn==0.23.0
scipy==1.4.1
six==1.14.0
sklearn==0.0
stevedore==1.32.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.6.0.post3
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==1.1.0
threadpoolctl==2.0.0
tifffile==2020.5.11
urllib3==1.25.9
virtualenv==20.0.21
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
